+++ src/bin/mix_ts.rs
@@ -47,14 +47,31 @@ async fn main() -> Result<(), Failure> {
 async fn run() -> Result<(), anyhow::Error> {
     let cfg = teamspeak::Config::new("ts3.ts3.online:8722")
         .channel("[cspacer]Best-of-Trance-Radio")
-        .name("🤖 ephyr::play_ts");
+        .name("🤖 ephyr::mix_ts");
     let mut ts_input = teamspeak::Input::new(cfg);
 
-    let ffmpeg = Command::new("ffplay")
+    let ffmpeg = Command::new("ffmpeg")
         .args(&["-loglevel", "debug"])
+        .args(&["-thread_queue_size", "512"])
+        .args(&["-i", "rtmp://127.0.0.1:1935/input/live"])
+        .args(&["-thread_queue_size", "512"])
         .args(&["-f", "f32be", "-sample_rate", "48000", "-channels", "2"])
         .args(&["-use_wallclock_as_timestamps", "true"])
         .args(&["-i", "pipe:0"])
+        .args(&[
+            "-filter_complex",
+            "[0:a]volume@org=0.5,\
+                  aresample=48000,\
+                  azmq=bind_address=tcp\\\\\\://0.0.0.0\\\\\\:6001[org];\
+             [1:a]volume@trn=1.5,\
+                  aresample=async=1,\
+                  azmq=bind_address=tcp\\\\\\://0.0.0.0\\\\\\:6002[trn];\
+             [org][trn]amix=inputs=2:duration=longest[out]",
+        ])
+        .args(&["-map", "[out]", "-map", "0:v"])
+        .args(&["-max_muxing_queue_size", "50000000"])
+        .args(&["-c:a", "libfdk_aac", "-c:v", "copy", "-shortest"])
+        .args(&["-f", "tee", "[f=flv]rtmp://127.0.0.1:1935/output/live"])
         .stdin(Stdio::piped())
         .stderr(Stdio::inherit())
         .kill_on_drop(true)