+++ src/filter/mod.rs
@@ -1,3 +1,0 @@
-//! Additional mixing filters.
-
-pub mod silence;