+++ src/util/mod.rs
@@ -1,6 +1,0 @@
-//! Helper utils.
-
-mod backoff;
-
-#[doc(inline)]
-pub use self::backoff::Backoff;